Move servlet tests to servlet folder and modify to use the ServletTestUtil class.
